+++ plugins/screenshot/helpers.js
